Cleaned up the code for DB access
