--- pg.js
@@ -1,48 +1,0 @@
-const pg = require('pg');
-
-/**
- * TODO(developer): specify SQL connection details
- */
-const connectionName =
-  process.env.INSTANCE_CONNECTION_NAME || 'plantit-296404:us-central1:plantit';
-const dbUser = process.env.SQL_USER || 'postgres';
-const dbPassword = process.env.SQL_PASSWORD || 'qwe123qwe';
-const dbName = process.env.SQL_NAME || 'postgres';
-
-const pgConfig = {
-  max: 1,
-  user: dbUser,
-  password: dbPassword,
-  database: dbName,
-};
-
-if (process.env.NODE_ENV === 'production') {
-  pgConfig.host = `/cloudsql/${connectionName}`;
-}
-
-// Connection pools reuse connections between invocations,
-// and handle dropped or expired connections automatically.
-let pgPool;
-
-exports.postgresDemo = (req, res) => {
-  // Initialize the pool lazily, in case SQL access isn't needed for this
-  // GCF instance. Doing so minimizes the number of active SQL connections,
-  // which helps keep your GCF instances under SQL connection limits.
-  if (!pgPool) {
-    pgPool = new pg.Pool(pgConfig);
-  }
-
-  pgPool.query('SELECT NOW() as now', (err, results) => {
-    if (err) {
-      console.error(err);
-      res.status(500).send(err);
-    } else {
-      res.send(JSON.stringify(results));
-    }
-  });
-
-  // Close any SQL resources that were declared inside this function.
-  // Keep any declared in global scope (e.g. mysqlPool) for later reuse.
-
-  
-};
